+++ src/bittytax/conv/parsers/__init__.py
@@ -39,6 +39,7 @@ from . import (
     cryptsy,
     deribit,
     easycrypto,
+    defitaxes_us,
     electrum,
     eternl,
     etherscan,