+++ src/bittytax/transactions.py
@@ -326,6 +326,7 @@ class TransactionBase:  # pylint: disable=too-many-instance-attributes
 class Buy(TransactionBase):  # pylint: disable=too-many-instance-attributes
     ACQUISITION_TYPES = {
         TrType.MINING,
+        TrType.STAKING_REWARD,
         TrType.STAKING,
         TrType.INTEREST,
         TrType.DIVIDEND,